Revert "clangsa: Refactor for increased precision (#61960)"
This reverts commit c2779f34f04859e914308b5b91147a3cc5b0fb02.
