+++ doc/src/devdocs/gc-sa.md
@@ -169,23 +169,16 @@ void example() {
 }
 ```
 
-### `JL_PROPAGATES_ROOT`/`JL_PROPAGATES_ROOT_INDEXED(root, index)`
+### `JL_PROPAGATES_ROOT`
 
 This annotation is commonly found on accessor functions that return one rootable
 object stored within another. When annotated on a function argument, it tells
 the analyzer that the root for that argument also applies to the value returned
 by the function.
-Use `JL_PROPAGATES_ROOT_INDEXED(root, index)` when the return value is loaded
-from a specific indexed child of the rooting argument, where `root` and `index`
-are zero-based argument indices. The indexed form lets the analyzer model later
-overwrites or clears of that indexed child precisely.
-Non-literal indices conservatively fall back to ordinary root propagation,
-because distinct symbolic index expressions can alias in the analyzer.
 
 Usage Example:
 ```c
-jl_value_t *jl_svecref(jl_svec_t *t JL_PROPAGATES_ROOT, size_t i)
-    JL_PROPAGATES_ROOT_INDEXED(0, 1) JL_NOTSAFEPOINT;
+jl_value_t *jl_svecref(jl_svec_t *t JL_PROPAGATES_ROOT, size_t i) JL_NOTSAFEPOINT;
 
 size_t example(jl_svec_t *svec) {
   jl_value_t *val = jl_svecref(svec, 1)
@@ -196,34 +189,22 @@ size_t example(jl_svec_t *svec) {
 }
 ```
 
-### `JL_ROOTED_BY_ARG(n)`/`JL_ROOTED_BY_ARG_INDEXED(root, index)`/`JL_OUT_ROOTED_BY_ARG(n)`/`JL_ROOTED_BY_RETURN`
+### `JL_ROOTING_ARGUMENT`/`JL_ROOTED_ARGUMENT`
 
-These are essentially the assignment counterpart to `JL_PROPAGATES_ROOT`.
+This is essentially the assignment counterpart to `JL_PROPAGATES_ROOT`.
 When assigning a value to a field of another value that is already rooted,
 the assigned value will inherit the root of the value it is assigned into.
 
-Use `JL_ROOTED_BY_ARG(n)` on the argument that is being assigned, where `n`
-is the zero-based argument index of the rooting argument. Use
-`JL_ROOTED_BY_ARG_INDEXED(root, index)` when the assigned value is stored in a
-specific indexed child of the rooting argument, where `index` is the zero-based
-argument index of the index argument. Use `JL_OUT_ROOTED_BY_ARG(n)` on an out
-argument when the value written through the out argument is rooted by argument
-`n`. Use `JL_ROOTED_BY_RETURN` on arguments that are rooted by the returned
-value. Variadic arguments cannot be annotated individually, so functions whose
-variadic arguments are rooted by the return value use `JL_ROOTED_VARARGS` on the
-function declaration.
-
 Usage Example:
 ```c
-void jl_svecset(void *t, size_t i, void *x JL_ROOTED_BY_ARG_INDEXED(0, 1)) JL_NOTSAFEPOINT;
-jl_svec_t *jl_svec1(void *a JL_ROOTED_BY_RETURN);
+void jl_svecset(void *t JL_ROOTING_ARGUMENT, size_t i, void *x JL_ROOTED_ARGUMENT) JL_NOTSAFEPOINT
 
 
 size_t example(jl_svec_t *svec) {
   jl_value_t *val = jl_box_long(10000);
-  jl_svecset(svec, 0, val);
-  // This is valid, because the annotation implies that jl_svecset
-  // propagates the rooted-ness from `svec` to `val`
+  jl_svecset(svec, val);
+  // This is valid, because the annotations imply that the
+  // jl_svecset propagates the rooted-ness from `svec` to `val`
   jl_gc_safepoint();
   return jl_unbox_long(val);
 }
