+++ src/method.c
@@ -743,7 +743,7 @@ JL_DLLEXPORT jl_code_instance_t *jl_cached_uninferred(jl_code_instance_t *codein
     return NULL;
 }
 
-JL_DLLEXPORT jl_code_instance_t *jl_cache_uninferred(jl_method_instance_t *mi JL_PROPAGATES_ROOT, jl_code_instance_t *checked, size_t world, jl_code_instance_t *newci)
+JL_DLLEXPORT jl_code_instance_t *jl_cache_uninferred(jl_method_instance_t *mi, jl_code_instance_t *checked, size_t world, jl_code_instance_t *newci)
 {
     while (!jl_mi_try_insert(mi, checked, newci)) {
         jl_code_instance_t *new_checked = jl_atomic_load_relaxed(&mi->cache);
@@ -759,7 +759,7 @@ JL_DLLEXPORT jl_code_instance_t *jl_cache_uninferred(jl_method_instance_t *mi JL
 
 // Return a newly allocated CodeInfo for the function signature
 // effectively described by the tuple (specTypes, env, Method) inside linfo
-JL_DLLEXPORT jl_code_info_t *jl_code_for_staged(jl_method_instance_t *mi JL_PROPAGATES_ROOT, size_t world, jl_code_instance_t **cache JL_OUT_ROOTED_BY_ARG(0))
+JL_DLLEXPORT jl_code_info_t *jl_code_for_staged(jl_method_instance_t *mi, size_t world, jl_code_instance_t **cache)
 {
     jl_code_instance_t *cache_ci = jl_atomic_load_relaxed(&mi->cache);
     jl_code_instance_t *uninferred_ci = jl_cached_uninferred(cache_ci, world);
@@ -1098,9 +1098,7 @@ JL_DLLEXPORT jl_method_t *jl_new_method_uninit(jl_module_t *module)
 // it will be the signature supplied in an `invoke` call.
 // If you don't need `invokesig`, you can set it to NULL on input.
 // Initialize iteration with `i = 0`. Returns `i` for the next backedge to be extracted.
-int get_next_edge(jl_array_t *list JL_PROPAGATES_ROOT, int i,
-                  jl_value_t **invokesig JL_OUT_ROOTED_BY_ARG(0),
-                  jl_code_instance_t **caller JL_OUT_ROOTED_BY_ARG(0)) JL_NOTSAFEPOINT
+int get_next_edge(jl_array_t *list, int i, jl_value_t** invokesig, jl_code_instance_t **caller) JL_NOTSAFEPOINT
 {
     jl_value_t *item = jl_array_ptr_ref(list, i);
     if (!item || jl_is_code_instance(item)) {
@@ -1120,9 +1118,7 @@ int get_next_edge(jl_array_t *list JL_PROPAGATES_ROOT, int i,
     return i + 2;
 }
 
-int set_next_edge(jl_array_t *list JL_PROPAGATES_ROOT, int i,
-                  jl_value_t *invokesig JL_ROOTED_BY_ARG(0),
-                  jl_code_instance_t *caller JL_ROOTED_BY_ARG(0)) JL_NOTSAFEPOINT
+int set_next_edge(jl_array_t *list, int i, jl_value_t *invokesig, jl_code_instance_t *caller)
 {
     if (invokesig)
         jl_array_ptr_set(list, i++, invokesig);
@@ -1130,8 +1126,7 @@ int set_next_edge(jl_array_t *list JL_PROPAGATES_ROOT, int i,
     return i;
 }
 
-int clear_next_edge(jl_array_t *list JL_PROPAGATES_ROOT, int i,
-                    jl_value_t *invokesig, jl_code_instance_t *caller) JL_NOTSAFEPOINT
+int clear_next_edge(jl_array_t *list, int i, jl_value_t *invokesig, jl_code_instance_t *caller)
 {
     if (invokesig)
         jl_array_ptr_set(list, i++, NULL);
